Fix compatibility with gcc-4.7
Many C++ headers don't pollute the namespace anymore with an implicit unistd.h inclusion,
so we need to include it ourselves in CardBase.cpp for sleep() prototype.

my_getopt prototypes don't match up with something else and cause an error, but
we really don't want to use my_getopt anyways - we were simply missing a
HAVE_GETOPT_H check with cmake for when system getopt.h is present.
Hardcode that setting for now - ideally cmake would make a check for it and
handle it automatically.